+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Follow Changelog
 
+## 1.0.2 - 2020-06-19
+### Changed
+- Updated documentation URL
+- `check` method now requires parameters. See documentation.
+
+### Added
+- `following` method can now be output as a string or array. This works well when trying to get a count for amount of followers & following.
+
+### Updated
+
 ## 1.0.1 - 2019-03-27
 ### Changed
 - Updated Icon